pyproject.toml
[project]
name = "coding_assistant_webui"
version = "0.1.0"
description = "A simple coding assistant app powered by Microsoft guidance, OpenAI LLMs, and Gradio."
authors = [
    { name = "discus0434", email = "discus0434@gmail.com" }
]
dependencies = ["guidance~=0.0.57", "openai~=0.27.7", "python-dotenv~=1.0.0", "gradio~=3.32.0", "setuptools~=67.8.0"]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["black~=23.3.0", "mypy~=1.3.0", "isort~=5.12.0"]
[tool.hatch.metadata]
allow-direct-references = true
